clang-tidy: effective config with warnings-as-errors (#537)
* clang-tidy: effective config with warnings-as-errors, fix all findings

Rework .clang-tidy from the stale CLion template into a high-signal config
and make the tree clean under it.

Config:
- enable bugprone-*, clang-analyzer-*, performance-*, portability-*
- disable convention-conflicting / high-noise checks (avoid-pragma-once,
  easily-swappable-parameters, narrowing-conversions, enum-size)
- WarningsAsErrors: '*' (clang-analyzer-* kept advisory: its only findings are
  inside third-party -isystem headers)
- drop the removed AnalyzeTemporaryDtors key that made clang-tidy reject the
  whole file on every run
- test/.clang-tidy disables unchecked-optional-access for tests, where gtest
  assertion macros defeat the dataflow analysis (false positives only)

Code fixes (all 148 source findings):
- performance-avoid-endl: std::endl -> '\n'
- bugprone-implicit-widening-of-multiplication-result: widen before multiply
- bugprone-forward-declaration-namespace: drop dead/wrong-namespace fwd decls
- bugprone-reserved-identifier: rename __ throwaways; NOLINT wvware's struct
- bugprone-empty-catch: NOLINT intentional best-effort catches
- bugprone-exception-escape: drop noexcept where functions can throw (string/
  path allocation); keep move ops noexcept; wrap CLI main() in try/catch
- bugprone-unchecked-optional-access: store re-fetched accessors, key ternaries
  on the optional, add a decryptor guard; NOLINT genuine invariants
- misc: inefficient-string-concatenation, unused-local, unsafe-functions
  (rewind -> fseek), optional-value-conversion, macro-parentheses, branch-clone

+++ src/odr/internal/cfb/cfb_impl.cpp
@@ -217,7 +217,8 @@ Sector CompoundFileReader::resolve_next_sector(std::istream &in,
   const std::uint32_t fatSectorLocation =
       resolve_fat_sector_location(in, fatSectorNumber);
   const std::uint64_t address = sector_offset_to_address(
-      {fatSectorLocation, sector % entriesPerSector * 4});
+      {fatSectorLocation,
+       static_cast<std::uint64_t>(sector % entriesPerSector) * 4});
   in.seekg(static_cast<std::streampos>(address));
   return parse_uint32(in);
 }
@@ -225,7 +226,8 @@ Sector CompoundFileReader::resolve_next_sector(std::istream &in,
 Sector CompoundFileReader::resolve_next_mini_sector(
     std::istream &in, const std::uint32_t mini_sector) const {
   const SectorOffset sector_offset = normalize_sector_offset(
-      in, {m_header.first_mini_fat_sector_location, mini_sector * 4});
+      in, {m_header.first_mini_fat_sector_location,
+           static_cast<std::uint64_t>(mini_sector) * 4});
   const std::uint64_t address = sector_offset_to_address(sector_offset);
   in.seekg(static_cast<std::streampos>(address));
   return parse_uint32(in);
@@ -294,8 +296,8 @@ std::uint32_t CompoundFileReader::resolve_fat_sector_location(
     in.seekg(static_cast<std::streampos>(address));
     difatSectorLocation = parse_uint32(in);
   }
-  const std::uint64_t address =
-      sector_offset_to_address({difatSectorLocation, fat_sector_number * 4});
+  const std::uint64_t address = sector_offset_to_address(
+      {difatSectorLocation, static_cast<std::uint64_t>(fat_sector_number) * 4});
   in.seekg(static_cast<std::streampos>(address));
   return parse_uint32(in);
 }
